# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bible
    description: NET Bible reader operations
  - name: reactions
    description: Article reaction and private reporting operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /bible/books:
    get:
      operationId: listBibleBooks
      tags: [bible]
      summary: List Bible books
      description: Returns the canonical book catalog used by the Bible reader.
      responses:
        "200":
          description: Bible books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BibleBook"
  /bible/passage:
    get:
      operationId: getBiblePassage
      tags: [bible]
      summary: Get a NET Bible passage
      description: Retrieves a passage from the official NET Bible service as safe plain text.
      parameters:
        - name: passage
          in: query
          required: true
          description: A human-readable Bible reference such as John 3 or Romans 8:1-11.
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        "200":
          description: NET Bible passage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiblePassage"
        "400":
          description: Invalid passage reference
        "404":
          description: Passage not found
        "502":
          description: Upstream Bible service unavailable
  /articles/{articleSlug}/reactions:
    get:
      operationId: getArticleReactions
      tags: [reactions]
      summary: Get public article reactions
      parameters:
        - $ref: "#/components/parameters/ArticleSlug"
      responses:
        "200":
          description: Public counts and the current visitor reaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleReactionSummary"
    put:
      operationId: setArticleReaction
      tags: [reactions]
      summary: Set or change the current visitor reaction
      parameters:
        - $ref: "#/components/parameters/ArticleSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleReactionInput"
      responses:
        "200":
          description: Updated public counts and current visitor reaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleReactionSummary"
        "400":
          description: Invalid article or reaction
  /reaction-admin/session:
    get:
      operationId: getReactionAdminSession
      tags: [reactions]
      summary: Check the private reaction report session
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionAdminSession"
    post:
      operationId: createReactionAdminSession
      tags: [reactions]
      summary: Sign in to the private reaction report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionAdminCredentials"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionAdminSession"
        "401":
          description: Incorrect password
        "503":
          description: Administrator password is not configured
    delete:
      operationId: deleteReactionAdminSession
      tags: [reactions]
      summary: Sign out of the private reaction report
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionAdminSession"
  /reaction-admin/stats:
    get:
      operationId: listReactionStats
      tags: [reactions]
      summary: List private article reaction totals
      responses:
        "200":
          description: Reaction totals by article
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleReactionStats"
        "401":
          description: A valid private report session is required
components:
  parameters:
    ArticleSlug:
      name: articleSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        minLength: 3
        maxLength: 160
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BibleBook:
      type: object
      required:
        - id
        - name
        - testament
        - chapters
      properties:
        id:
          type: string
          example: john
        name:
          type: string
          example: John
        testament:
          type: string
          enum: [Old Testament, New Testament]
        chapters:
          type: number
          minimum: 1
          example: 21
    BibleVerse:
      type: object
      required:
        - bookName
        - chapter
        - verse
        - text
      properties:
        bookName:
          type: string
          example: John
        chapter:
          type: number
          minimum: 1
          example: 3
        verse:
          type: number
          minimum: 1
          example: 16
        text:
          type: string
        type:
          type: string
          example: verse
    BiblePassage:
      type: object
      required:
        - reference
        - version
        - verses
        - copyright
      properties:
        reference:
          type: string
          example: John 3
        version:
          type: string
          example: NET
        verses:
          type: array
          items:
            $ref: "#/components/schemas/BibleVerse"
        copyright:
          type: string
          description: Required NET Bible copyright acknowledgement.
    ArticleReactionType:
      type: string
      enum: [helpful, encouraging, disagree]
    ArticleReactionInput:
      type: object
      required: [reaction]
      properties:
        reaction:
          $ref: "#/components/schemas/ArticleReactionType"
    ArticleReactionSummary:
      type: object
      required: [articleSlug, helpful, encouraging, selected]
      properties:
        articleSlug:
          type: string
        helpful:
          type: ["number", "null"]
          minimum: 5
          description: Public total, or null until it reaches five.
        encouraging:
          type: ["number", "null"]
          minimum: 5
          description: Public total, or null until it reaches five.
        selected:
          anyOf:
            - $ref: "#/components/schemas/ArticleReactionType"
            - type: "null"
    ReactionAdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 200
    ReactionAdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
    ArticleReactionStats:
      type: object
      required: [articleSlug, helpful, encouraging, disagree, total]
      properties:
        articleSlug:
          type: string
        helpful:
          type: number
          minimum: 0
        encouraging:
          type: number
          minimum: 0
        disagree:
          type: number
          minimum: 0
        total:
          type: number
          minimum: 0
        updatedAt:
          type: ["string", "null"]
          format: date-time