# lib/api-zod/src/generated/types/articleReactionStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleReactionStats {
  articleSlug: string;
  /** @minimum 0 */
  helpful: number;
  /** @minimum 0 */
  encouraging: number;
  /** @minimum 0 */
  disagree: number;
  /** @minimum 0 */
  total: number;
  /** @nullable */
  updatedAt?: Date | null;
}