# lib/api-zod/src/generated/types/reactionAdminCredentials.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionAdminCredentials {
  /**
     * @minLength 1
     * @maxLength 200
     */
  password: string;
}